requirements.txt
# --- CORE: The Brain ---
langgraph>=0.0.10
langchain>=0.1.0
langchain-community>=0.0.10
langchain-nvidia-ai-endpoints>=0.0.1
python-dotenv>=1.0.0

# --- NOLA: The Voice (Modern Stack) ---
vosk>=0.3.45          # Offline Speech Recognition
sounddevice>=0.4.6    # Audio Input/Output
edge-tts>=6.1.0       # Microsoft Edge Neural TTS
pygame>=2.5.0         # Audio Playback
pycaw>=2024.0         # Modern Audio Control
comtypes>=1.2.0       # Required for Pycaw

# --- IRIS: The Vision ---
mss>=9.0.0            # Fast Screen Capture
Pillow>=10.0.0        # Image Processing
winsdk>=1.0.0         # Windows OCR for Sentry

# --- TARA: The Tools ---
psutil>=5.9.0         # System Monitoring
AppOpener>=1.7        # App Launching
pyperclip>=1.8.0      # Clipboard
pywhatkit>=5.4        # YouTube/Web automation
duckduckgo-search>=6.0 # Web Search

# --- UI/CLI ---
prompt_toolkit>=3.0.0
numpy>=1.23.0
requests>=2.28.0

# --- Dev/Test (Optional) ---
pytest>=7.0
